feat: acoes de listar, emprestar e devolver livros

+++ biblioteca/src/Autor.java
@@ -2,21 +2,21 @@ import java.util.Date;
 
 public class Autor {
 
-    Long id;
+    int id;
     String nome;
     Date dataNascimento;
 
-    public Autor(Long id, String nome, Date dataNascimento) {
+    public Autor(int id, String nome, Date dataNascimento) {
         this.id = id;
         this.nome = nome;
         this.dataNascimento = dataNascimento;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
